🔨 [packages] Move class `PackageRepositoryLoader` to `loader` module
